pyproject.toml
[project]
name = "artefactual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hicham Randrianarivo", email = "hicham.randrianarivo@gmail.com" },
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "absl-py>=2.1.0",
    "beartype>=0.19.0",
    "etils[eapp,edc,enp,epath,epy,etqdm,etree]>=1.11.0",
    "keras-hub>=0.18.1",
    "keras>3",
    "scikit-learn>=1.6.0",
    "tensorflow-datasets>=4.9.7",
    "jax[cuda12]; platform_system != 'Darwin'",
    "jax; platform_system == 'Darwin'",
    "clu>=0.0.12",
    "tensorflow-cpu<2.18; platform_system != 'Darwin'",
    "tensorflow<2.18; platform_system == 'Darwin'",
    "transformers>=4.50.3",
]

[project.scripts]
artefactual = "artefactual:main"

[project.optional-dependencies]
hub = ["huggingface-hub>=0.27.0", "keras-hub>=0.18.1", "safetensors>=0.4.5"]
test = ["hypothesis>=6.127.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "datasets>=3.2.0",
    "einops>=0.8.0",
    "faiss-cpu>=1.9.0.post1",
    "fastembed>=0.4.2 ; sys_platform == 'darwin'",
    "grain>=0.2.3",
    "ipython>=8.31.0",
    "jinja2>=3.1.5",
    "jupytext>=1.16.6",
    "scikeras>=0.13.0",
    "toolz>=1.0.0",
    "wandb>=0.19.4",
    "lm-eval[vllm]>=0.4.7 ; sys_platform != 'darwin'",
    "mlcroissant>=1.0.12",
    "polars>=1.21.0",
    "orjson>=3.10.15",
    "tokenizers>=0.21.0",
    "transformers>=4.48.1",
    "outlines>=0.1.11",
    "pudb>=2024.1.3",
    "pydantic>=2.10.6",
    "pytype>=2024.10.11",
    "jupyterlab>=4.3.5",
    "hypothesis>=6.127.8",
]

[tool.codespell]
ignore-words-list = "rouge,ans"

[tool.ruff]
preview = true
target-version = "py310"
line-length = 120

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]

ignore = ["ISC001", "F811"]

[tool.ruff.lint.per-file-ignores]
"**/scripts/*" = [
    "INP001",
    "T201",
    "PLR0917",
    "PLR0913",
    "RUF009",
    "FBT003",
    "E501",
    "PLR0914",
    "A004",
]
"**/tests/**/*" = ["PLC1901", "PLR2004", "PLR6301", "S", "TID252"]
"docs/notebooks/**/*" = ["INP001", "T201", "A004", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["argimi"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.jupytext]
cell_markers = '"""'

[tool.pytype]
# Space-separated list of files or directories to exclude.
exclude = ["**/*_test.py", "**/test_*.py"]

# Space-separated list of files or directories to process.
inputs = ["."]

# Keep going past errors to analyze as many files as possible.
keep_going = false

# Run N jobs in parallel. When 'auto' is used, this will be equivalent to the
# number of CPUs on the host system.
jobs = 4

# All pytype output goes here.
output = ".pytype"

# Bind 'self' in methods with non-transparent decorators. This flag is temporary
# and will be removed once this behavior is enabled by default.
bind_decorated_methods = true

# Don't allow None to match bool. This flag is temporary and will be removed
# once this behavior is enabled by default.
none_is_not_bool = true

# Enable parameter count checks for overriding methods with renamed arguments.
# This flag is temporary and will be removed once this behavior is enabled by
# default.
overriding_renamed_parameter_count_checks = true

# Variables initialized as None retain their None binding. This flag is
# temporary and will be removed once this behavior is enabled by default.
strict_none_binding = false

# Support the third-party fiddle library. This flag is temporary and will be
# removed once this behavior is enabled by default.
use_fiddle_overlay = false

# Opt-in: Do not allow Any as a return type.
no_return_any = false

# Opt-in: Require decoration with @typing.override when overriding a method or
# nested class attribute of a parent class.
require_override_decorator = true

# Experimental: Infer precise return types even for invalid function calls.
precise_return = true

# Experimental: Solve unknown types to label with structural types.
# Enabling this seems to disable typechecking
protocols = false

# Experimental: Only load submodules that are explicitly imported.
strict_import = true

# Experimental: Enable exhaustive checking of function parameter types.
strict_parameter_checks = true

# Experimental: Emit errors for comparisons between incompatible primitive
# types.
strict_primitive_comparisons = true

# Experimental: Check that variables are defined in all possible code paths.
strict_undefined_checks = true

# Experimental: FOR TESTING ONLY. Use pytype/rewrite/.
use_rewrite = false

# Space-separated list of error names to ignore.
disable = ["pyi-error"]

# Don't report errors.
report_errors = true

[tool.uv]
dev-dependencies = [
    "ruff>=0.14.8",
    "ty>=0.0.1a32",
]
